docker-compose/app/requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
mysqlclient==2.2.1
packaging==23.2
protobuf==4.21.12
Werkzeug==3.0.1
gunicorn==21.2.0
prometheus-client==0.19.0
psutil==5.9.6
